test-requirements.txt
pytest == 7.*
pytest-cases == 3.8.5
pytest-drop-dup-tests
pytest-ordering
pytest-html
pytest-json-report
pytest-dependency
jinja2
bcrypt
requests
paramiko
pycryptodome
sshpubkeys
pyyaml
polling2
cryptography
boto3
scp
requests
urllib3
lxml
pyyaml
pdoc3
tox
openstacksdk == 4.1.0
dataclasses; python_version < '3.7'
-e ./apiclient
